move WS things around
